Revert "Feature/refactor modules"
