+++ package.json
@@ -28,6 +28,8 @@
     "@nestjs/config": "^2.2.0",
     "@nestjs/core": "^9.0.0",
     "@nestjs/platform-express": "^9.0.0",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.13.2",
     "ethers": "^5.7.1",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
