+++ src/controllers/__tests__/proxy.controller.spec.ts
@@ -3,14 +3,12 @@ import { ConfigService } from '@nestjs/config';
 import { ForbiddenException } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
 import { Response } from 'express';
+import { TransactionService, VerseService, ProxyService } from 'src/services';
+import { ProxyController } from '../proxy.controller';
 import {
-  TransactionService,
-  VerseService,
-  ProxyService,
   AllowCheckService,
   JsonrpcCheckService,
-} from 'src/services';
-import { ProxyController } from 'src/controllers';
+} from 'src/shared/services/src';
 
 describe('ProxyController', () => {
   let jsonrpcCheckService: JsonrpcCheckService;
