+++ src/services/index.ts
@@ -1,5 +1,3 @@
 export * from './proxy.service';
 export * from './transaction.service';
 export * from './verse.service';
-export * from './jsonrpcCheck.service';
-export * from './allowCheck.service';
