+++ src/shared/entities/Error.ts
