+++ src/shared/entities/Verse.ts
