+++ src/shared/entities/index.ts
