+++ src/shared/services/service.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { AllowCheckService, JsonrpcCheckService } from './src';
+
+@Module({
+  providers: [AllowCheckService, JsonrpcCheckService],
+})
+export class ServiceModule {}
