+++ src/shared/services/src/__tests__/jsonrpcCheck.service.spec.ts
@@ -1,4 +1,4 @@
-import { JsonrpcCheckService } from 'src/services';
+import { JsonrpcCheckService } from '../jsonrpcCheck.service';
 
 const jsonrpcCheckService = new JsonrpcCheckService();
