+++ src/shared/services/src/allowCheck.service.ts
