+++ src/shared/services/src/index.ts
@@ -1,0 +1,2 @@
+export * from './allowCheck.service';
+export * from './jsonrpcCheck.service';
