+++ src/shared/shared.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { ServiceModule } from './services/service.module';
+
+@Module({
+  imports: [ServiceModule],
+  exports: [ServiceModule],
+})
+export class SharedModule {}
